ai_modules/rmbg_service/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
onnxruntime==1.17.1
opencv-python-headless==4.9.0.80
Pillow==10.2.0
numpy==1.26.4
python-multipart==0.0.9
pydantic==2.6.1
